Add functionality to send notifications to users
Passes stage 5
This update brings changes to DataManager, User, and Main classes. A new method was added to DataManager to retrieve users that need to be notified when they reach maximum points for a course. Some modifications have also been implemented in the User class to store a list of completed and notified courses. In the Main class, a new command has been established to utilize these enhancements and send notifications to the appropriate users.